pyproject.toml
[tool.poetry]
name = "Team-11-Project"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
pymysql = "^1.1.0"
cryptography = "^42.0.7"
loguru = "^0.7.2"
pydantic-core = "^2.18.2"
annotated-types = "^0.6.0"
h11 = "^0.14.0"
paho-mqtt = "^2.1.0"
httpx = "^0.27.0"
dashscope = "^1.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"